% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mithril-package.R
\docType{package}
\name{mithril-package}
\alias{mithril-package}
\alias{mithril}
\title{mithril: A Personal Collection of R Functions}
\description{
The mithril package provides a set of utility functions developed for
personal use, covering various aspects of data analysis and visualization.
Currently, it includes a function to plot confusion matrices.
}
\section{Key Functions}{

\itemize{
\item \code{\link{plot_conf_matrix}}: Generates a heatmap of a confusion matrix.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RobWiederstein/mithril}
  \item \url{https://robwiederstein.github.io/mithril/}
  \item Report bugs at \url{https://github.com/RobWiederstein/mithril/issues}
}

}
\author{
\strong{Maintainer}: Rob Wiederstein \email{khuon68@gmail.com}

}
\keyword{internal}
