% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mithril-package.R
\docType{package}
\name{mithril-package}
\alias{mithril-package}
\alias{mithril}
\title{mithril: A Personal Collection of R Functions}
\description{
The mithril package provides a set of utility functions developed for
personal use, covering various aspects of data analysis and visualization.
Currently, it includes a function to plot confusion matrices.
}
\section{Key Functions}{

\itemize{
\item \code{\link{plot_conf_matrix}}: Generates a heatmap of a confusion matrix.
}
}

\author{
\strong{Maintainer}: Rob Wiederstein \email{khuon68@gmail.com}

}
\keyword{internal}
